made it so no duplicates can be added to localstorage, also started seperating components
