+++ package.json
@@ -13,8 +13,6 @@
   "dependencies": {
     "express": "^4.18.1",
     "node-fetch": "^2.6.7",
-    "pg": "^8.8.0",
-    "react-router-dom": "^6.3.0",
     "sequelize": "^6.21.4"
   },
   "devDependencies": {
